Remove duplicate Nav bars
This moves the main navigation bar to the frontend App.js and passes all
session information down to the Frontpage via props.
